# Cached string decoder for map keys (CachedKeyDecoder)