Check this out @gfx We are using it in our project.

Big part of all strings are map keys and most of the time total number of uniq keys is small.
For example in our payload in 260k objects 67% of all string are keys with 100 uniq values.
I've tried to cache this values and skip string decoding at all for keys.
It's using bytes representation of key string and stores actual key value. So instead of decoding key string we just need to find this key value in cache.

Results before:
```
Benchmark on NodeJS/v12.3.1

operation                                                         |   op   |   ms  |  op/s 
----------------------------------------------------------------- | ------: | ----: | ------:
buf = Buffer.from(JSON.stringify(obj));                           |  557200 |  5000 |  111440
buf = JSON.stringify(obj);                                        | 1078100 |  5000 |  215620
obj = JSON.parse(buf);                                            |  394300 |  5001 |   78844
buf = require("msgpack-lite").encode(obj);                        |  416400 |  5000 |   83280
obj = require("msgpack-lite").decode(buf);                        |  313600 |  5000 |   62720
buf = require("@msgpack/msgpack").encode(obj);                    |  646100 |  5000 |  129220
obj = require("@msgpack/msgpack").decode(buf);                    |  561800 |  5000 |  112360
✨  Done in 36.69s.

```

After:
```
Benchmark on NodeJS/v12.3.1

operation                                                         |   op   |   ms  |  op/s 
----------------------------------------------------------------- | ------: | ----: | ------:
buf = Buffer.from(JSON.stringify(obj));                           |  598900 |  5000 |  119780
buf = JSON.stringify(obj);                                        | 1151000 |  5000 |  230200
obj = JSON.parse(buf);                                            |  397600 |  5000 |   79520
buf = require("msgpack-lite").encode(obj);                        |  429000 |  5000 |   85800
obj = require("msgpack-lite").decode(buf);                        |  323800 |  5000 |   64760
buf = require("@msgpack/msgpack").encode(obj);                    |  666800 |  5000 |  133360
obj = require("@msgpack/msgpack").decode(buf);                    |  744000 |  5000 |  148800
✨  Done in 36.95s.
```

I am not sure that we should include it directly into this library, but maybe it can be added as optional speedup that user can add if it's required.

It works better if Decoder is reused, because cache is need to be created first.

Let me know what do you think about it.