@gfx 2019-06-03
Fascinating idea! I'll merge the string caching in decoders, but there are some improvement and changes needed.


## The reuse of Decoder

If one would like to decode it in the fastest way, they should reuse the instance of `Decoder`, anyway. We should re-write the benchmark script to reuse `Decoder`.

And, I think `decode()` should create `Decoder` for each time in order to control configuration easily.

## Default behavior

I'd like to use caching by default, so it should be merged to Decoder.

Configuration of `cacheSize: number` may be a good idea to control memory usage because uncontrolled caching is a kind of memory leak.