@sergeyzenchenko 2019-06-03
Sure, this is just a prototype solution that we use, it's not intended to be merged directly. It requires changes and tests.

I think we can allow user to pass string decoder to options, so user can pass custom decoder if needed. 

What do you think about it?