@sergeyzenchenko 2019-06-03
I am not sure about how to properly implement cacheSize yet.
In worst case scenario cacheSize can be filled with low frequently used keys if they are frequently used in first documents.

Maybe I can add ```cache.compact()``` function that user can call on demand.
Or this can be called periodically inside Decoder.
This function will remove less frequently used keys.

Current version tracks number of hits for each key and orders them by this number to make most frequently used keys faster.