@sergeyzenchenko 2019-06-03
Lol @gfx , I've just improved performance more :D Now it's 176660 ops/s