@gfx 2019-06-04
> Lol @gfx , I've just improved performance more :D Now it's 176660 ops/s

Amazing! 😆 

---

cacheSize is not necessarily correct; it's just a hint.

---

> I think we can allow user to pass string decoder to options, so user can pass custom decoder if needed

You mean `decode(buffer, { stringDecoder: sharedCachingStringDecoder })`? I prefer `{ stringCache: sharedStringCache }` because its intent is clearer so users can focus on the cache engine, not the whole string decoder.

---

BTW I have another idea to share  encoders and decoders, not directly related to this PR: https://github.com/msgpack/msgpack-javascript/issues/56