@gfx 2019-07-26
I'd like to proceed this patch this week. FYI:  Chrome 76 is going to be released at the end of this month, which includes JSON.parse performance improvement.