@gfx 2019-08-02
```
Benchmark on NodeJS/v12.7.0

operation                                                         |   op   |   ms  |  op/s
----------------------------------------------------------------- | ------: | ----: | ------:
buf = Buffer.from(JSON.stringify(obj));                           |  507700 |  5000 |  101540
buf = JSON.stringify(obj);                                        |  958100 |  5000 |  191620
obj = JSON.parse(buf);                                            |  346500 |  5000 |   69300
buf = require("msgpack-lite").encode(obj);                        |  361800 |  5001 |   72345
obj = require("msgpack-lite").decode(buf);                        |  267400 |  5000 |   53480
buf = require("@msgpack/msgpack").encode(obj);                    |  510200 |  5000 |  102040
obj = require("@msgpack/msgpack").decode(buf);                    |  825500 |  5000 |  165100
```